Fix nil error in MerchantProduct price helpers
When creating a new product, default_price_cents is nil, causing the
default_price_dollars method to fail with "undefined method '/' for nil"
when the form attempts to display the field value.

Fixed by:
- Adding nil check to default_price_dollars (returns nil if price not set)
- Adding nil check to price_with_unit (returns "Not set" if price not set)

This allows the new product form to render properly for products without
a price set yet.